requirements.txt
# Gemini-RAG Dependencies
# Python 3.12+ required

# Core dependencies for File Search Store demo
google-genai>=1.49.0
python-dotenv>=1.0.0
pydantic>=2.0.0
